refactor: encapsulate bot
