+++ src/bot/commands/addrepo.ts
@@ -1,10 +1,10 @@
 import { Command } from "@grammyjs/commands";
 
-import type { BotContext } from "@/bot";
-
 import { config } from "@/config";
 import { db, schema } from "@/db";
 import { gitHubRepoName, isGitHubUrl } from "@/github";
+
+import type { BotContext } from "../bot";
 
 // TODO: Fetch and store contributors too
 export async function addrepoHandler(ctx: BotContext) {
