+++ src/bot/commands/removerepo.ts
@@ -1,13 +1,13 @@
 import { Command } from "@grammyjs/commands";
 import { eq } from "drizzle-orm";
 
-import type { BotContext } from "@/bot";
-
 import { config } from "@/config";
 import { db, schema } from "@/db";
 import { isGitHubUrl } from "@/github";
 
-import { escapeMarkdown } from "../lib/escape-markdown";
+import type { BotContext } from "../bot";
+
+import { escapeMarkdown } from "../../lib/escape-markdown";
 
 export async function removerepoHandler(ctx: BotContext) {
   if (!ctx.message) return;
