+++ src/bot/commands/whoami.ts
@@ -1,12 +1,12 @@
 import { Command } from "@grammyjs/commands";
 import { eq } from "drizzle-orm";
 
-import type { BotContext } from "@/bot";
-
 import { config } from "@/config";
 import { db, schema } from "@/db";
 
-import { escapeMarkdown } from "../lib/escape-markdown";
+import type { BotContext } from "../bot";
+
+import { escapeMarkdown } from "../../lib/escape-markdown";
 
 export async function whoamiHandler(ctx: BotContext) {
   if (!ctx.message) return; // the update must be a new message.
